+++ src/About/About.js
@@ -1,0 +1,10 @@
+import React, { Component } from 'react';
+
+
+const About = () => {
+    return (
+        <p>About Page</p>
+    )
+}
+
+export default About;
