+++ src/Affirmations/PowerAffirmations/PowerAffirmations.scss
@@ -1,0 +1,188 @@
+$desktop: 840px;
+@mixin desktop {
+  @media (min-width: #{$desktop}){
+      @content;
+  }
+}
+  /*Wrapper*/
+  .wrapper{
+    display: grid;
+    grid-template-columns: 1;
+    grid-template-rows: 20% auto;
+    grid-template-areas:
+    "sentence"
+    "fadeIn";
+    font-family: 'Raleway', sans-serif;
+    margin: 100px auto;
+    padding: 40px 40px;
+    position: relative;
+    width: 70%;
+    @include desktop {
+      grid-template-columns: 30% auto;
+      grid-template-rows: 1;
+      grid-template-areas:
+      "sentence fadeIn"
+    }
+    /*Sentence*/
+    .sentence {
+      grid-area: sentence;
+      justify-self: center;
+     color: white;
+     font-size: 30px;
+     text-align: center;
+     position: relative;
+    // left: -340px;
+    @include desktop {
+     
+      
+    }
+     
+  /*FadeIn*/
+  .fadeIn{
+    grid-area: fadeIn;
+    position: absolute;
+    top: 50px;
+    left: 75px;
+    @include desktop {
+      top: -50px;
+      left: 300px;
+      justify-self: center;
+      position: relative;
+    }
+  }
+  //Animation, add 1.5 seconds on for each nth-child and another 1.5 seconds for total.
+    .fadeIn span{
+    animation: fadeEffect 29s linear infinite 0s;
+    color: #00abe9;
+    opacity: 0;
+    overflow: hidden;
+    position: absolute;
+    }
+    .fadeIn span:nth-child(2){
+  animation-delay: 1s;
+  
+  
+    }
+    .fadeIn span:nth-child(3){
+      animation-delay: 2s;
+      
+     
+    }
+    .fadeIn span:nth-child(4){
+      animation-delay: 3s;
+      
+      
+    }
+    .fadeIn span:nth-child(5){
+      animation-delay: 4s;
+      
+      
+    }
+    .fadeIn span:nth-child(6){
+        animation-delay: 5s;
+        
+      }
+      .fadeIn span:nth-child(7){
+        animation-delay: 6s;
+        
+      }
+      .fadeIn span:nth-child(8){
+        animation-delay: 7s;
+        
+      }
+      .fadeIn span:nth-child(9){
+        animation-delay: 8s;
+        
+      }
+      .fadeIn span:nth-child(10){
+        animation-delay: 9s;
+        
+      }
+      .fadeIn span:nth-child(11){
+        animation-delay: 10s;
+        
+      }
+      .fadeIn span:nth-child(12){
+        animation-delay: 11s;
+        
+      }
+      .fadeIn span:nth-child(13){
+        animation-delay: 12s;
+        
+      }
+      .fadeIn span:nth-child(14){
+        animation-delay: 13s;
+        
+      }
+      .fadeIn span:nth-child(15){
+        animation-delay: 14s;
+        
+      }
+      .fadeIn span:nth-child(16){
+        animation-delay: 15s;
+        
+      }
+      .fadeIn span:nth-child(17){
+        animation-delay: 16s;
+        
+      }
+      .fadeIn span:nth-child(18){
+        animation-delay: 17s;
+        
+      }
+      .fadeIn span:nth-child(19){
+        animation-delay: 18s;
+        
+      }
+      .fadeIn span:nth-child(20){
+        animation-delay: 19s;
+        
+      }
+      .fadeIn span:nth-child(21){
+        animation-delay: 20s;
+        
+      }
+      .fadeIn span:nth-child(22){
+        animation-delay: 21s;
+        
+      }
+      .fadeIn span:nth-child(23){
+        animation-delay: 22s;
+        
+      }
+      .fadeIn span:nth-child(24){
+        animation-delay: 23s;
+        
+      }
+      .fadeIn span:nth-child(25){
+        animation-delay: 24s;
+        
+      }
+      .fadeIn span:nth-child(26){
+        animation-delay: 25s;
+        
+      }
+      .fadeIn span:nth-child(27){
+        animation-delay: 26s;
+        
+      }
+      .fadeIn span:nth-child(28){
+        animation-delay: 27s;
+        
+      }
+      .fadeIn span:nth-child(29){
+        animation-delay: 28s;
+        
+      }
+      
+    
+  }
+  }
+  /*Scale Animation*/
+  @keyframes fadeEffect{
+  0% { opacity: 0; }
+  1% { opacity: 0; transform: scale(1); }
+  3% { opacity: 1; transform: scale(3); }
+  4% { opacity: 0; transform: scale(0); }
+  100% { opacity: 0; }
+  }
