+++ src/App.js
@@ -1,36 +1,27 @@
 import React, { Component } from 'react';
 import logo from './logo.svg';
 import './CSS/App.css';
+import { BrowserRouter as Router, Route, Link } from "react-router-dom";
+import Header from './Header/Header';
+import About from './About/About';
+import PowerAffirmations from './Affirmations/PowerAffirmations/PowerAffirmations';
+import AllAffirmations from './Affirmations/AllAffirmations';
 
 class App extends Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      words: ["I", "am", "a", "positive", "person", "[loading...]",
-              "I", "have", "many", "unique", "skills", "[loading...]",
-              "I", "will", "use", "these", "skills", "effectively", "[loading...]",
-              "I", "look", "at", "all", "situations", "in", "a", "positive", "light", "[loading...]"]
-    }
-  }
   
   render() {
-    const words = this.state.words.map((word) => {
-      return <span key={word}>{word}</span>;
-     });
+    
     
     return (
       <div className="App">
-        <header className="App-header">
-        <h1>Empowering Affirmations</h1>
-          <section className="wrapper">
-            <h2 className="sentence">Say to yourself:
-              <div className="fadeIn">
-                {words}
-
-              </div>
-            </h2>
-          </section>
-        </header>
+        <Router>
+        <Header />
+          <header className="AppContainer">            
+            <Route exact path="/" component={AllAffirmations} />
+            <Route exact path="/PowerAffirmations" component={PowerAffirmations} />
+            <Route path="/about" component={About} />
+          </header>
+        </Router>
       </div>
     );
   }
