+++ src/CSS/App.css
@@ -1,10 +1,10 @@
 .App {
   text-align: center;
+  background-color: #282c34;
 }
 
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
+.AppContainer {
+  height: 100vh;
   display: -webkit-box;
   display: -ms-flexbox;
   display: flex;
@@ -22,330 +22,5 @@
   color: White;
 }
 
-/*Wrapper*/
-.wrapper {
-  display: -ms-grid;
-  display: grid;
-  -ms-grid-columns: 1;
-      grid-template-columns: 1;
-  -ms-grid-rows: 20% auto;
-      grid-template-rows: 20% auto;
-      grid-template-areas: "sentence" "fadeIn";
-  font-family: 'Raleway', sans-serif;
-  margin: 100px auto;
-  padding: 40px 40px;
-  position: relative;
-  width: 70%;
-  /*Sentence*/
-}
-
-@media (min-width: 840px) {
-  .wrapper {
-    -ms-grid-columns: 30% auto;
-        grid-template-columns: 30% auto;
-    -ms-grid-rows: 1;
-        grid-template-rows: 1;
-        grid-template-areas: "sentence fadeIn";
-  }
-}
-
-.wrapper .sentence {
-  -ms-grid-row: 1;
-  -ms-grid-column: 1;
-  grid-area: sentence;
-  -ms-grid-column-align: center;
-      justify-self: center;
-  color: white;
-  font-size: 30px;
-  text-align: center;
-  position: relative;
-  /*FadeIn*/
-  /*Pop Effect*/
-}
-
-.wrapper .sentence .fadeIn {
-  -ms-grid-row: 2;
-  -ms-grid-column: 1;
-  grid-area: fadeIn;
-  position: absolute;
-  top: 50px;
-  left: 75px;
-}
-
-@media (min-width: 840px) {
-  .wrapper .sentence .fadeIn {
-    top: -50px;
-    left: 300px;
-    -ms-grid-column-align: center;
-        justify-self: center;
-    position: relative;
-  }
-}
-
-.wrapper .sentence .fadeIn span {
-  -webkit-animation: fadeEffect 29s linear infinite 0s;
-          animation: fadeEffect 29s linear infinite 0s;
-  color: #00abe9;
-  opacity: 0;
-  overflow: hidden;
-  position: absolute;
-}
-
-.wrapper .sentence .fadeIn span:nth-child(2) {
-  -webkit-animation-delay: 1s;
-          animation-delay: 1s;
-}
-
-.wrapper .sentence .fadeIn span:nth-child(3) {
-  -webkit-animation-delay: 2s;
-          animation-delay: 2s;
-}
-
-.wrapper .sentence .fadeIn span:nth-child(4) {
-  -webkit-animation-delay: 3s;
-          animation-delay: 3s;
-}
-
-.wrapper .sentence .fadeIn span:nth-child(5) {
-  -webkit-animation-delay: 4s;
-          animation-delay: 4s;
-}
-
-.wrapper .sentence .fadeIn span:nth-child(6) {
-  -webkit-animation-delay: 5s;
-          animation-delay: 5s;
-}
-
-.wrapper .sentence .fadeIn span:nth-child(7) {
-  -webkit-animation-delay: 6s;
-          animation-delay: 6s;
-}
-
-.wrapper .sentence .fadeIn span:nth-child(8) {
-  -webkit-animation-delay: 7s;
-          animation-delay: 7s;
-}
-
-.wrapper .sentence .fadeIn span:nth-child(9) {
-  -webkit-animation-delay: 8s;
-          animation-delay: 8s;
-}
-
-.wrapper .sentence .fadeIn span:nth-child(10) {
-  -webkit-animation-delay: 9s;
-          animation-delay: 9s;
-}
-
-.wrapper .sentence .fadeIn span:nth-child(11) {
-  -webkit-animation-delay: 10s;
-          animation-delay: 10s;
-}
-
-.wrapper .sentence .fadeIn span:nth-child(12) {
-  -webkit-animation-delay: 11s;
-          animation-delay: 11s;
-}
-
-.wrapper .sentence .fadeIn span:nth-child(13) {
-  -webkit-animation-delay: 12s;
-          animation-delay: 12s;
-}
-
-.wrapper .sentence .fadeIn span:nth-child(14) {
-  -webkit-animation-delay: 13s;
-          animation-delay: 13s;
-}
-
-.wrapper .sentence .fadeIn span:nth-child(15) {
-  -webkit-animation-delay: 14s;
-          animation-delay: 14s;
-}
-
-.wrapper .sentence .fadeIn span:nth-child(16) {
-  -webkit-animation-delay: 15s;
-          animation-delay: 15s;
-}
-
-.wrapper .sentence .fadeIn span:nth-child(17) {
-  -webkit-animation-delay: 16s;
-          animation-delay: 16s;
-}
-
-.wrapper .sentence .fadeIn span:nth-child(18) {
-  -webkit-animation-delay: 17s;
-          animation-delay: 17s;
-}
-
-.wrapper .sentence .fadeIn span:nth-child(19) {
-  -webkit-animation-delay: 18s;
-          animation-delay: 18s;
-}
-
-.wrapper .sentence .fadeIn span:nth-child(20) {
-  -webkit-animation-delay: 19s;
-          animation-delay: 19s;
-}
-
-.wrapper .sentence .fadeIn span:nth-child(21) {
-  -webkit-animation-delay: 20s;
-          animation-delay: 20s;
-}
-
-.wrapper .sentence .fadeIn span:nth-child(22) {
-  -webkit-animation-delay: 21s;
-          animation-delay: 21s;
-}
-
-.wrapper .sentence .fadeIn span:nth-child(23) {
-  -webkit-animation-delay: 22s;
-          animation-delay: 22s;
-}
-
-.wrapper .sentence .fadeIn span:nth-child(24) {
-  -webkit-animation-delay: 23s;
-          animation-delay: 23s;
-}
-
-.wrapper .sentence .fadeIn span:nth-child(25) {
-  -webkit-animation-delay: 24s;
-          animation-delay: 24s;
-}
-
-.wrapper .sentence .fadeIn span:nth-child(26) {
-  -webkit-animation-delay: 25s;
-          animation-delay: 25s;
-}
-
-.wrapper .sentence .fadeIn span:nth-child(27) {
-  -webkit-animation-delay: 26s;
-          animation-delay: 26s;
-}
-
-.wrapper .sentence .fadeIn span:nth-child(28) {
-  -webkit-animation-delay: 27s;
-          animation-delay: 27s;
-}
-
-.wrapper .sentence .fadeIn span:nth-child(29) {
-  -webkit-animation-delay: 28s;
-          animation-delay: 28s;
-}
-
-.wrapper .sentence .popEffect {
-  display: inline;
-  text-indent: 8px;
-}
-
-.wrapper .sentence .popEffect span {
-  animation: pop 12.5s linear infinite 0s;
-  -ms-animation: pop 12.5s linear infinite 0s;
-  -webkit-animation: pop 12.5s linear infinite 0s;
-  color: #00abe9;
-  opacity: 0;
-  overflow: hidden;
-  position: absolute;
-}
-
-.wrapper .sentence .popEffect span:nth-child(2) {
-  animation-delay: 2.5s;
-  -ms-animation-delay: 2.5s;
-  -webkit-animation-delay: 2.5s;
-}
-
-.wrapper .sentence .popEffect span:nth-child(3) {
-  animation-delay: 5s;
-  -ms-animation-delay: 5s;
-  -webkit-animation-delay: 5s;
-}
-
-.wrapper .sentence .popEffect span:nth-child(4) {
-  animation-delay: 7.5s;
-  -ms-animation-delay: 7.5s;
-  -webkit-animation-delay: 7.5s;
-}
-
-.wrapper .sentence .popEffect span:nth-child(5) {
-  animation-delay: 10s;
-  -ms-animation-delay: 10s;
-  -webkit-animation-delay: 10s;
-}
-
-/*Scale Animation*/
-@-webkit-keyframes fadeEffect {
-  0% {
-    opacity: 0;
-  }
-  1% {
-    opacity: 0;
-    -webkit-transform: scale(1);
-            transform: scale(1);
-  }
-  3% {
-    opacity: 1;
-    -webkit-transform: scale(3);
-            transform: scale(3);
-  }
-  4% {
-    opacity: 0;
-    -webkit-transform: scale(0);
-            transform: scale(0);
-  }
-  100% {
-    opacity: 0;
-  }
-}
-@keyframes fadeEffect {
-  0% {
-    opacity: 0;
-  }
-  1% {
-    opacity: 0;
-    -webkit-transform: scale(1);
-            transform: scale(1);
-  }
-  3% {
-    opacity: 1;
-    -webkit-transform: scale(3);
-            transform: scale(3);
-  }
-  4% {
-    opacity: 0;
-    -webkit-transform: scale(0);
-            transform: scale(0);
-  }
-  100% {
-    opacity: 0;
-  }
-}
-
 /*FadeIn Animation*/
-/*Pop Effect Animation*/
-
-@-webkit-keyframes pop {
-  0% {
-    opacity: 0;
-  }
-  5% {
-    opacity: 0;
-    -webkit-transform: rotate(0deg) scale(0.1) skew(0deg) translate(0px);
-  }
-  10% {
-    opacity: 1;
-    -webkit-transform: translateY(0px);
-  }
-  25% {
-    opacity: 1;
-    -webkit-transform: translateY(0px);
-  }
-  30% {
-    opacity: 0;
-    -webkit-transform: translateY(0px);
-  }
-  80% {
-    opacity: 0;
-  }
-  100% {
-    opacity: 0;
-  }
-}
 /*# sourceMappingURL=App.css.map */
