+++ src/CSS/App.scss
@@ -1,239 +1,19 @@
-$desktop: 840px;
-@mixin desktop {
-  @media (min-width: #{$desktop}){
-      @content;
-  }
-}
+
 .App {
     text-align: center;
+    background-color: #282c34;
   }
   
   
-  .App-header {
-    background-color: #282c34;
-    min-height: 100vh;
+  .AppContainer {
+    
+    height: 100vh;
     display: flex;
     flex-direction: column;
     align-items: center;
     justify-content: center;
     font-size: calc(10px + 2vmin);
     color: White;
-  }
-  /*Wrapper*/
-  .wrapper{
-    display: grid;
-    grid-template-columns: 1;
-    grid-template-rows: 20% auto;
-    grid-template-areas:
-    "sentence"
-    "fadeIn";
-    font-family: 'Raleway', sans-serif;
-    margin: 100px auto;
-    padding: 40px 40px;
-    position: relative;
-    width: 70%;
-    @include desktop {
-      grid-template-columns: 30% auto;
-      grid-template-rows: 1;
-      grid-template-areas:
-      "sentence fadeIn"
-    }
-    /*Sentence*/
-    .sentence {
-      grid-area: sentence;
-      justify-self: center;
-     color: white;
-     font-size: 30px;
-     text-align: center;
-     position: relative;
-    // left: -340px;
-    @include desktop {
-     
-      
-    }
-     
-  /*FadeIn*/
-  .fadeIn{
-    grid-area: fadeIn;
-    position: absolute;
-    top: 50px;
-    left: 75px;
-    @include desktop {
-      top: -50px;
-      left: 300px;
-      justify-self: center;
-      position: relative;
-    }
-  }
-  //Animation, add 1.5 seconds on for each nth-child and another 1.5 seconds for total.
-    .fadeIn span{
-    animation: fadeEffect 29s linear infinite 0s;
-    color: #00abe9;
-    opacity: 0;
-    overflow: hidden;
-    position: absolute;
-    }
-    .fadeIn span:nth-child(2){
-  animation-delay: 1s;
-  
-  
-    }
-    .fadeIn span:nth-child(3){
-      animation-delay: 2s;
-      
-     
-    }
-    .fadeIn span:nth-child(4){
-      animation-delay: 3s;
-      
-      
-    }
-    .fadeIn span:nth-child(5){
-      animation-delay: 4s;
-      
-      
-    }
-    .fadeIn span:nth-child(6){
-        animation-delay: 5s;
-        
-      }
-      .fadeIn span:nth-child(7){
-        animation-delay: 6s;
-        
-      }
-      .fadeIn span:nth-child(8){
-        animation-delay: 7s;
-        
-      }
-      .fadeIn span:nth-child(9){
-        animation-delay: 8s;
-        
-      }
-      .fadeIn span:nth-child(10){
-        animation-delay: 9s;
-        
-      }
-      .fadeIn span:nth-child(11){
-        animation-delay: 10s;
-        
-      }
-      .fadeIn span:nth-child(12){
-        animation-delay: 11s;
-        
-      }
-      .fadeIn span:nth-child(13){
-        animation-delay: 12s;
-        
-      }
-      .fadeIn span:nth-child(14){
-        animation-delay: 13s;
-        
-      }
-      .fadeIn span:nth-child(15){
-        animation-delay: 14s;
-        
-      }
-      .fadeIn span:nth-child(16){
-        animation-delay: 15s;
-        
-      }
-      .fadeIn span:nth-child(17){
-        animation-delay: 16s;
-        
-      }
-      .fadeIn span:nth-child(18){
-        animation-delay: 17s;
-        
-      }
-      .fadeIn span:nth-child(19){
-        animation-delay: 18s;
-        
-      }
-      .fadeIn span:nth-child(20){
-        animation-delay: 19s;
-        
-      }
-      .fadeIn span:nth-child(21){
-        animation-delay: 20s;
-        
-      }
-      .fadeIn span:nth-child(22){
-        animation-delay: 21s;
-        
-      }
-      .fadeIn span:nth-child(23){
-        animation-delay: 22s;
-        
-      }
-      .fadeIn span:nth-child(24){
-        animation-delay: 23s;
-        
-      }
-      .fadeIn span:nth-child(25){
-        animation-delay: 24s;
-        
-      }
-      .fadeIn span:nth-child(26){
-        animation-delay: 25s;
-        
-      }
-      .fadeIn span:nth-child(27){
-        animation-delay: 26s;
-        
-      }
-      .fadeIn span:nth-child(28){
-        animation-delay: 27s;
-        
-      }
-      .fadeIn span:nth-child(29){
-        animation-delay: 28s;
-        
-      }
-      
-    
-    /*Pop Effect*/
-    .popEffect{
-      display: inline;
-      text-indent: 8px;
-    }
-    .popEffect span{
-      animation: pop 12.5s linear infinite 0s;
-      -ms-animation: pop 12.5s linear infinite 0s;
-      -webkit-animation: pop 12.5s linear infinite 0s;
-      color: #00abe9;
-      opacity: 0;
-      overflow: hidden;
-      position: absolute;
-    }
-    .popEffect span:nth-child(2){
-      animation-delay: 2.5s;
-      -ms-animation-delay: 2.5s;
-      -webkit-animation-delay: 2.5s;
-    }
-    .popEffect span:nth-child(3){
-      animation-delay: 5s;
-      -ms-animation-delay: 5s;
-      -webkit-animation-delay: 5s;
-    }
-    .popEffect span:nth-child(4){
-      animation-delay: 7.5s;
-      -ms-animation-delay: 7.5s;
-      -webkit-animation-delay: 7.5s;
-    }
-    .popEffect span:nth-child(5){
-      animation-delay: 10s;
-      -ms-animation-delay: 10s;
-      -webkit-animation-delay: 10s;
-    }
-  }
-  }
-  /*Scale Animation*/
-  @keyframes fadeEffect{
-  0% { opacity: 0; }
-  1% { opacity: 0; transform: scale(1); }
-  3% { opacity: 1; transform: scale(3); }
-  4% { opacity: 0; transform: scale(0); }
-  100% { opacity: 0; }
   }
   
   
@@ -269,33 +49,5 @@ $desktop: 840px;
   // }
   
   
-  /*Pop Effect Animation*/
-  @-moz-keyframes pop{
-  0% { opacity: 0; }
-  5% { opacity: 0; -moz-transform: rotate(0deg) scale(0.10) skew(0deg) translate(0px); }
-  10% { opacity: 1; -moz-transform: translateY(0px); }
-  25% { opacity: 1; -moz-transform: translateY(0px); }
-  30% { opacity: 0; -moz-transform: translateY(0px); }
-  80% { opacity: 0; }
-  100% { opacity: 0;}
-  }
-  @-webkit-keyframes pop{
-  0% { opacity: 0; }
-  5% { opacity: 0; -webkit-transform: rotate(0deg) scale(0.10) skew(0deg) translate(0px);}
-  10% { opacity: 1; -webkit-transform: translateY(0px); }
-  25% { opacity: 1; -webkit-transform: translateY(0px); }
-  30% { opacity: 0; -webkit-transform: translateY(0px); }
-  80% { opacity: 0; }
-  100% { opacity: 0; }
-  }
-  @-ms-keyframes pop{
-  0% { opacity: 0; }
-  5% { opacity: 0; -ms-transform: rotate(0deg) scale(0.10) skew(0deg) translate(0px); }
-  10% { opacity: 1; -ms-transform: translateY(0px); }
-  25% { opacity: 1; -ms-transform: translateY(0px); }
-  30% { opacity: 0; -ms-transform: translateY(0px); }
-  80% { opacity: 0; }
-  100% { opacity: 0; }
-  }
