+++ src/Header/Header.scss
@@ -1,0 +1,20 @@
+.Header {
+    
+
+    padding: 20px;
+    .linkTitles {
+        font-size: 2em;
+        font-weight: bold;
+    }
+   
+}
+.Header > span {
+    margin: 100px;
+}
+a {
+    color: white;
+    text-decoration: none;
+    &:hover {
+        color: orange;
+    }
+}
